+++ packages/components/ui/src/sensors/index.js
@@ -1,1 +1,2 @@
 export {_screen, default as ScreenSensor} from './screen/ScreenSensor.svelte';
+export {default as StyleSensor} from './style/StyleSensor.svelte';