+++ packages/tools/utils/src/modules/string-[string-array].js
@@ -5,6 +5,7 @@
 import * as _ from 'lamb';
 
 import {split} from './string_proto-array.js';
+import {trim} from './string_proto-string.js';
 
 /**
  * Return a function expecting a separator or regex to split the provided string
@@ -43,3 +44,28 @@ export const makeSplitStringBy = _.curry(split, 2);
  * @since 0.1.0
  */
 export const makeSplitBy = _.curryRight(split, 2);
+
+/**
+ * Return a function that splits the expected string and trims all the elements
+ * of the returned array
+ * @see {@link module:@svizzle/utils/string-[string-array].makeSplitBy|makeSplitBy}
+ *
+ * @function
+ * @arg {string} separator
+ * @return {function} - String -> Array
+ *
+ * @example
+> trimSplitByDoubleDot = makeTrimmedSplitBy('..')
+>
+> trimSplitByDoubleDot('  aa ..\ta\n..a')
+['aa', 'a', 'a']
+>
+> trimSplitByDoubleDot('  aa ...\na..a.a.aa\n.....\taa..\n')
+['aa', '.\na', 'a.a.aa', '', '.\taa', '']
+ *
+ * @since 0.19.0
+ */
+export const makeTrimmedSplitBy = separator => _.pipe([
+	_.splitBy(separator),
+	_.mapWith(trim)
+]);